requirements.txt
selenium
pytest
pytest-xdist==2.3.0
webdriver-manager==3.3.0
allure-pytest==2.8.18





